# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SarvamGuideRequest {
  /**
     * @minLength 2
     * @maxLength 300
     */
  question: string;
}

export interface SarvamGuideResponse {
  answer: string;
  /**
     * @minItems 2
     * @maxItems 3
     */
  suggestions: string[];
}

export interface SarvamSpeechRequest {
  /**
     * @minLength 1
     * @maxLength 600
     */
  text: string;
}

export type SarvamSpeechResponseMimeType = typeof SarvamSpeechResponseMimeType[keyof typeof SarvamSpeechResponseMimeType];


export const SarvamSpeechResponseMimeType = {
  'audio/wav': 'audio/wav',
} as const;

export interface SarvamSpeechResponse {
  audioBase64: string;
  mimeType: SarvamSpeechResponseMimeType;
}

export type SarvamTranscriptionRequestMimeType = typeof SarvamTranscriptionRequestMimeType[keyof typeof SarvamTranscriptionRequestMimeType];


export const SarvamTranscriptionRequestMimeType = {
  'audio/webm': 'audio/webm',
  'audio/ogg': 'audio/ogg',
  'audio/wav': 'audio/wav',
  'audio/mp4': 'audio/mp4',
  'audio/mpeg': 'audio/mpeg',
} as const;

export interface SarvamTranscriptionRequest {
  /**
     * Short-lived base64 audio payload; never persisted by the server.
     * @minLength 100
     * @maxLength 2400000
     */
  audioBase64: string;
  mimeType: SarvamTranscriptionRequestMimeType;
}

export interface SarvamTranscriptionResponse {
  /**
     * @minLength 2
     * @maxLength 300
     */
  transcript: string;
}

export interface ApiError {
  error: string;
}