# lib/api-zod/src/generated/types/sarvamGuideResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SarvamGuideResponse {
  answer: string;
  /**
     * @minItems 2
     * @maxItems 3
     */
  suggestions: string[];
}