# lib/api-zod/src/generated/types/sarvamSpeechResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SarvamSpeechResponseMimeType } from './sarvamSpeechResponseMimeType';

export interface SarvamSpeechResponse {
  audioBase64: string;
  mimeType: SarvamSpeechResponseMimeType;
}