# lib/api-zod/src/generated/types/sarvamTranscriptionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SarvamTranscriptionRequestMimeType } from './sarvamTranscriptionRequestMimeType';

export interface SarvamTranscriptionRequest {
  /**
     * Short-lived base64 audio payload; never persisted by the server.
     * @minLength 100
     * @maxLength 2400000
     */
  audioBase64: string;
  mimeType: SarvamTranscriptionRequestMimeType;
}