# lib/api-zod/src/generated/types/sarvamTranscriptionResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SarvamTranscriptionResponse {
  /**
     * @minLength 2
     * @maxLength 300
     */
  transcript: string;
}